Refactored and moved current clients into their own folders.
