+++ RepoCleanup/Application/CommandHandlers/CreateDefaultTeamsCommandHandler.cs
@@ -1,6 +1,6 @@
 ﻿using RepoCleanup.Application.Commands;
 using RepoCleanup.Models;
-using RepoCleanup.Services;
+using RepoCleanup.Infrastructure.Clients.Gitea;
 using RepoCleanup.Utils;
 using System.Collections.Generic;
 using System.Threading.Tasks;
