+++ RepoCleanup/Functions/AddTeamToRepoFunction.cs
@@ -1,7 +1,6 @@
 ﻿using RepoCleanup.Application.CommandHandlers;
 using RepoCleanup.Application.Commands;
-using RepoCleanup.Models;
-using RepoCleanup.Services;
+using RepoCleanup.Infrastructure.Clients.Gitea;
 using System;
 using System.Collections.Generic;
 using System.Linq;
