+++ RepoCleanup/Infrastructure/Clients/Altinn2/Altinn2Form.cs
@@ -1,4 +1,4 @@
-﻿namespace RepoCleanup.Models
+﻿namespace RepoCleanup.Infrastructure.Clients.Altinn2
 {
     public class Altinn2Form
     {
