+++ RepoCleanup/Infrastructure/Clients/Altinn2/Altinn2Service.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace RepoCleanup.Models
+namespace RepoCleanup.Infrastructure.Clients.Altinn2
 {
     public class Altinn2Service
     {
