+++ RepoCleanup/Infrastructure/Clients/Gitea/CreateTeamOption.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace RepoCleanup.Models
+namespace RepoCleanup.Infrastructure.Clients.Gitea
 {
     public enum Permission
     {
