+++ RepoCleanup/Infrastructure/Clients/Gitea/Organisation.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace RepoCleanup.Models
+namespace RepoCleanup.Infrastructure.Clients.Gitea
 {
     public class Organisation
     {
