+++ RepoCleanup/Infrastructure/Clients/Gitea/TrustModel.cs
@@ -1,4 +1,4 @@
-﻿namespace RepoCleanup.Models
+﻿namespace RepoCleanup.Infrastructure.Clients.Gitea
 {
     public enum TrustModel
     {
