+++ RepoCleanup/RepoCleanup.csproj
@@ -23,4 +23,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Application\Models\" />
+  </ItemGroup>
+
 </Project>
